service/api/pyproject.toml
[project]
name = "astro-reason-api"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.30",
  "pydantic>=2.8",
  "confluent-kafka>=2.13",
  "psycopg[binary]>=3.2",
  "boto3>=1.35",
  "python-multipart>=0.0.9",
  "tenacity>=8.4",
]

[tool.uvicorn]
factory = false
host = "0.0.0.0"
port = 8000
